pyproject.toml
[project]
name = "repo-clarifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.82.0",
    "pyyaml>=6.0.2",
]
